Changed RGB Colorspace to floats to avoid rounding issues

+++ samples/Color/ColorExtractor/ColorSpaces/RGBColor.cs
@@ -2,17 +2,17 @@
 {
     public struct RGBColor
     {
-        public RGBColor(byte r, byte g, byte b)
+        public RGBColor(float r, float g, float b)
         {
             R = r;
             G = g;
             B = b;
         }
 
-        public byte R { get; set; }
+        public float R { get; set; }
 
-        public byte G { get; set; }
+        public float G { get; set; }
 
-        public byte B { get; set; }
+        public float B { get; set; }
     }
 }
